requirements.txt
python-telegram-bot[webhooks]==20.7
python-dotenv==1.0.0
openai>=1.0.0
httpx>=0.27.0
eth-account>=0.13.0
